services/user-service/requirements.txt
# --- Web Framework ---
fastapi==0.111.0
uvicorn[standard]==0.30.1

# --- OpenTelemetry (OTel) ---
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0

# --- OTel Instrumentation ---
# We only need the FastAPI instrumentation
opentelemetry-instrumentation-fastapi

# --- OTel Exporter ---
opentelemetry-exporter-otlp-proto-grpc==1.25.0
